# openapi/bss-mvp-api-v1.yaml
openapi: 3.1.0
info:
  title: BSS MVP API
  version: 1.1.0
  summary: Implemented Backend MVP contract for the frozen BSS Frontend v1.0.0.
  description: |
    Implemented versioned contract for the functional Backend MVP; deployment remains a separate
    operational decision. The server derives organization,
    role and department scope from the verified session/device identity. Client-supplied
    scope is never authorization evidence. CSV and XLSX are the currently approved export
    formats. PDF is approved as a professional business export in v1.1.
  x-bss-status: MVP_IMPLEMENTED
  x-bss-contract-reviewed-at: 2026-07-17
  x-bss-frontend-release: frontend-v1.0.0
  x-bss-baseline-commit: 91323c7cdbbbbf7b965c4926c94a11af6d31bf62
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Same-origin application API
...
security:
  - sessionCookie: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session and current user context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionContext'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      summary: Refresh Session
      security: []
      responses:
        '204':
          description: Session cookie rotated.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/invitations/accept:
    post:
      tags: [Auth]
      operationId: acceptInvitation
      summary: Accept Invitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptRequest'
      responses:
        '200':
          description: Invitation accepted, password set and secure session created atomically.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionContext'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Logout
      x-bss-roles: [admin, manager, worker, accountant]
      responses:
        '204':
          description: Session invalidated when present and both cookies cleared; repeating logout is safe.
        '403':
          $ref: '#/components/responses/Forbidden'
  /me:
    get:
      tags: [Auth]
      operationId: getSessionContext
      summary: Get Session Context
      x-bss-roles: [admin, manager, worker, accountant]
      responses:
        '200':
          description: Verified user, organization and effective scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionContext'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bss_session
      description: Secure, HttpOnly, SameSite session cookie.
    deviceSignature:
      type: apiKey
      in: header
      name: X-BSS-Signature
      description: |
        HMAC-SHA256 v1. The terminal signs UTF-8 bytes of
        METHOD + "\n" + PATH + "\n" + SHA256_HEX(BODY) + "\n" + X-BSS-Timestamp + "\n" + X-BSS-Nonce.
        The signature is lowercase hexadecimal. Server resolves X-BSS-Device-Id inside the tenant-owned
        encrypted credential store, permits at most 300 seconds clock skew and atomically consumes each nonce.
...
  responses:
    Unauthenticated:
      description: Missing, invalid or expired identity.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated identity is outside the required role or data scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationFailed:
      description: Stable domain or field validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Role:
      type: string
      enum: [admin, manager, worker, accountant]
...
    EntityStatus:
      type: string
      enum: [active, blocked]
    Revision:
      type: string
      pattern: '^[1-9][0-9]{0,18}$'
      description: Positive PostgreSQL bigint optimistic-concurrency revision.
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]+$'
        message:
          type: string
        requestId:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 1024
    InvitationAcceptRequest:
      type: object
      additionalProperties: false
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 256
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 1024
    SessionContext:
      type: object
      additionalProperties: false
      required: [user, organization, effectiveScope]
      properties:
        user:
          $ref: '#/components/schemas/User'
        organization:
          $ref: '#/components/schemas/Organization'
        effectiveScope:
          type: object
          required: [role, departmentIds, selfWorkerId]
          properties:
            role:
              $ref: '#/components/schemas/Role'
            departmentIds:
              type: array
              items:
                type: string
                format: uuid
            selfWorkerId:
              type: [string, 'null']
              format: uuid
...
    ApprovedLeaveVisibility:
      type: string
      enum: [team, department, organization]
    Organization:
      type: object
      required: [id, name, timezone, approvedLeaveVisibility, revision]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        taxIdentifier:
          type: string
        timezone:
          type: string
          examples: [Europe/Zagreb]
        approvedLeaveVisibility:
          $ref: '#/components/schemas/ApprovedLeaveVisibility'
        revision:
          $ref: '#/components/schemas/Revision'
...
    User:
      type: object
      required: [id, email, role, status, departmentIds, revision]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/EntityStatus'
        workerId:
          type: [string, 'null']
          format: uuid
        departmentIds:
          type: array
          items:
            type: string
            format: uuid
        revision:
          $ref: '#/components/schemas/Revision'
...